move noblegarden gameobject script to new script file noblegarden.cpp
