+++ src/server/scripts/Events/events_script_loader.cpp
@@ -24,6 +24,7 @@ void AddSC_event_love_in_the_air();
 void AddSC_event_midsummer_scripts();
 void AddSC_event_childrens_week();
 void AddSC_event_firework_show_scripts();
+void AddSC_event_noblegarden_scripts();
 
 // The name of this function should match:
 // void Add${NameOfDirectory}Scripts()
@@ -37,4 +38,5 @@ void AddEventsScripts()
     AddSC_event_midsummer_scripts();
     AddSC_event_childrens_week();
     AddSC_event_firework_show_scripts();
+    AddSC_event_noblegarden_scripts();
 }
